Event loop public api (#691)
