+++ source/linux/epoll_event_loop.c
@@ -3,17 +3,16 @@
  * SPDX-License-Identifier: Apache-2.0.
  */
 
-#include <aws/io/event_loop.h>
-
 #include <aws/cal/cal.h>
 #include <aws/common/atomics.h>
 #include <aws/common/clock.h>
 #include <aws/common/mutex.h>
 #include <aws/common/task_scheduler.h>
 #include <aws/common/thread.h>
-#include <aws/io/private/tracing.h>
-
+#include <aws/io/event_loop.h>
 #include <aws/io/logging.h>
+#include <aws/io/private/event_loop_impl.h>
+#include <aws/io/private/tracing.h>
 
 #include <sys/epoll.h>
