+++ source/posix/socket.c
@@ -13,6 +13,7 @@
 
 #include <aws/io/event_loop.h>
 #include <aws/io/logging.h>
+#include <aws/io/private/event_loop_impl.h>
 
 #include <arpa/inet.h>
 #include <aws/io/io.h>
