+++ source/windows/iocp/pipe.c
@@ -7,6 +7,7 @@
 
 #include <aws/common/task_scheduler.h>
 #include <aws/io/event_loop.h>
+#include <aws/io/private/event_loop_impl.h>
 
 #include <stdbool.h>
 #include <stdio.h>
