+++ tests/alpn_handler_test.c
@@ -5,6 +5,7 @@
 
 #include <aws/io/channel.h>
 #include <aws/io/event_loop.h>
+#include <aws/io/private/event_loop_impl.h>
 #include <aws/io/tls_channel_handler.h>
 #include <aws/testing/aws_test_harness.h>
